Use `Atomic = Transparent_atomic`, `fenceless_get`, and `incr`
Transparent atomic provides a workaround for the OCaml 5.(0|1) CSE bugs around
`Atomic.get`s.  In this case, invalid CSE optimization on `Atomic.get`s can
actually break the queue semantics.

Fences are not really needed in this specific queue with the exception of the
updates of the head and tail indices and even there the fences are only needed
to ensure that the other side sees the index update only after updates to the
array.

`Atomic.incr` is potentially a little bit faster, because it needs no write
barrier.