Avoid wide pattern matches and assume CSE instead
Wide pattern matches, even with immutable fields, tend to generate inefficient
code in OCaml at the moment.  Specifically, the compiler tends to generate a
sequence of loads for the matched fields at the start of the function.  The
problem with this is that it increases register pressure, which may lead to
spilling (i.e. extra reads and writes to stack), and may also lead to fetching
some fields unnecessarily (i.e. wasted work) as those fields are only accessed
conditionally.

It is typically better to access immutable fields using dot notation.  The
compiler performs CSE on such loads so repeated accesses of immutable fields do
not necessarily generate repeated loads of those fields.