Implement caching of the other side index
This reduces contention as operations can be performed on either side in
parallel without interference from the other side.

Without this algorithmic change, most of the other optimizations become
irrelevant, because the contention on index accesses would simply limit the peak
performance by an order of magnitude or more.  Try removing this change and
observe the massive drop in peak performance.