Avoid indirection in array and drop `mask` field
The `mask` can be computed very quickly from the length of the array which is
known to be a power of two.

Avoiding the indirection means that an allocation, some pointer chasing, and
potentially a couple of cache misses (an extra object needs to be written by the
producer (a potential cache miss) and read by consumer (a definite cache miss,
because the memory is written by the producer)) are avoided per message.  This
provides a significant performance improvement now that contention has been
reduced.

Using `Obj.magic ()` with an array that may contain `float` values is prone to
be problematic and it makes sense to test that things work.