Avoid float array pessimization
OCaml currently has a special case for `float array`.  This means than whenever
an array is accessed that might be a `float array`, the compiler generates an
extra conditional to treat the array properly.  These conditionals slow down
array accesses slightly.  Here we hack around that by specifying that the array
is of elements of a "pointer" type.  This way the compiler will generate array
accesses with potential write barriers and avoid generating code for the `float
array` special case.  This improves performance.