Avoid duplication using GADTs
This reduces code size at the expense of one or two conditionals.  The time
taken by those conditionals is fairly insignificant.  The reduction in code size
will likely improve performance in real applications, because pressure on the
instruction cache is reduced.

This approach is likely to be useful in various other data structure
implementations and is likely often preferable to inlining to avoid code
duplication.