Move node rep to another file, remove snapshot mechanism, tweak algo
This moves the node representation to a separate file so that it is easy
to change only the node representation.

This also changes the queue representation to be closer to the traditional
Michael-Scott queue where the head points to a node whose value is no longer
logically part of the queue.  The "snapshot" mechanism cannot work with this
representation, because the nodes must be mutated as they are logically removed
from the queue to free the values to avoid space leaks.

This commit also implements various micro-optimizations and tweaks to the
implementation that improve performance.  For example, backoffs are added after
failed CAS operations, which improves performance in case of contention.