Fix some typos, English idioms, and reformat the doc comments
As a stylistic thing, I would recommend using an extra space after a period.
Like this.  And this.  This makes it easier to see where sentences begin and end
when using monospaced fonts and makes the text easier to read.  The convention
of an extra space after a period is also recognized by the `fill-paragraph`
command of Emacs.